+++ assertions/src/main/java/internal/su/pernova/assertions/matchers/IsBoolean.java
@@ -21,6 +21,6 @@ public class IsBoolean extends DescriptiveMatcher {
 
 	@Override
 	public Description describe(Description description) {
-		return super.describe(description).appendArgument(expected);
+		return super.describe(description).appendExpected(expected);
 	}
 }