+++ assertions/src/main/java/internal/su/pernova/assertions/matchers/IsChar.java
@@ -25,6 +25,6 @@ public class IsChar extends DescriptiveMatcher {
 
 	@Override
 	public Description describe(Description description) {
-		return super.describe(description).appendArgument(expected);
+		return super.describe(description).appendExpected(expected);
 	}
 }