+++ assertions/src/main/java/internal/su/pernova/assertions/matchers/IsFloat.java
@@ -23,6 +23,6 @@ public class IsFloat extends DescriptiveMatcher {
 
 	@Override
 	public Description describe(Description description) {
-		return super.describe(description).appendArgument(expected);
+		return super.describe(description).appendExpected(expected);
 	}
 }