+++ assertions/src/main/java/internal/su/pernova/assertions/matchers/IsLong.java
@@ -24,6 +24,6 @@ public class IsLong extends DescriptiveMatcher {
 
 	@Override
 	public Description describe(Description description) {
-		return super.describe(description).appendArgument(expected);
+		return super.describe(description).appendExpected(expected);
 	}
 }